more qreal

+++ src/common/ColorThemeWorker.cpp
@@ -90,9 +90,9 @@ ColorThemeWorker::ColorThemeWorker(QObject *parent) : QObject (parent)
 
 QColor ColorThemeWorker::mergeColors(const QColor& upper, const QColor& lower) const
 {
-    float r1, g1, b1, a1;
-    float r2, g2, b2, a2;
-    float r, g, b, a;
+    qreal r1, g1, b1, a1;
+    qreal r2, g2, b2, a2;
+    qreal r, g, b, a;
 
     upper.getRgbF(&r1, &g1, &b1, &a1);
     lower.getRgbF(&r2, &g2, &b2, &a2);
